Removed the unique_together constraint on the Customer Model
For a given stripe_Account (djstripe_owner_account) and livemode, the same user (SUBSCRIBER_MODEL) can have multiple Customer Accounts. In fact Stripe's default behaviour is to create new accounts even for the same email. Hence the requirement that (subcriber, djstripe_owner_account, livemode) be unique is incorrect.

The only reason this wasn't caught until now is because the Customer was not getting associated correctly with the SUBSCRIBER_MODEL.